feat(pretrade): enforce margin limits
Problem:
- Futures and perpetual swap margin checks were still placeholders.
- The file-backed limit format could not carry static v1 margin requirements.
- Position::Option did not match the specification's deferred f64 delta shape.

Changes:
- Add checked notional-by-absolute-quantity multiplication for margin math.
- Implement initial-margin checks for futures and perpetual swaps.
- Add available margin to EvaluateRequest and static initial_margin_per_unit limits.
- Extend file limit parsing with initial_margin_per_unit records.
- Align Position::Option delta with the spec's Option<f64> placeholder.

Verification:
- cargo fmt --all --check passed.
- cargo test --workspace passed.
- cargo clippy --workspace --all-targets -- -D warnings passed.
- RUSTDOCFLAGS="-D warnings" cargo doc --workspace --no-deps passed.

Notes:
- Margin remains static per instrument for v1, consistent with the no live margin-schedule ingestion boundary.

+++ risk-core/src/position.rs
@@ -16,7 +16,7 @@ pub struct MarginState {
 }
 
 /// Dynamic position state, separate from static instrument reference data.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Position {
     /// Cash equity position.
     Equity {
@@ -68,7 +68,7 @@ pub enum Position {
         /// Signed position quantity.
         qty: Qty,
         /// Optional externally supplied delta, unused by v1 pretrade.
-        delta: Option<i64>,
+        delta: Option<f64>,
         /// Contract expiry.
         expiry: Timestamp,
     },
